Address Android Rust platform review
