+++ .github/workflows/ci.yml
@@ -56,6 +56,8 @@ jobs:
             runner: macos-latest
           - mise_env: macos-arm64-egl
             runner: macos-latest
+          - mise_env: android-arm64-vulkan
+            runner: macos-latest
           - mise_env: windows-x64-vulkan
             runner: windows-2022
           - mise_env: windows-x64-wgl
@@ -65,44 +67,51 @@ jobs:
       - uses: ./.github/actions/setup-ci-deps
 
       - run: mise run configure
-      - run: mise run test
+      - name: Build native library
+        if: ${{ startsWith(matrix.mise_env, 'android-') }}
+        run: mise run build
+      - name: Test native library
+        if: ${{ !startsWith(matrix.mise_env, 'android-') }}
+        run: mise run test
 
       - name: Build Java FFM binding
-        if: ${{ matrix.mise_env != 'macos-arm64-egl' }}
+        if: ${{ matrix.mise_env != 'macos-arm64-egl' && !startsWith(matrix.mise_env, 'android-') }}
         run: mise run //bindings/java-ffm:build
       - name: Build Java JNI binding
-        if: ${{ matrix.mise_env != 'macos-arm64-egl' }}
+        if: ${{ matrix.mise_env != 'macos-arm64-egl' && !startsWith(matrix.mise_env, 'android-') }}
         run: mise run //bindings/java-jni:build
       - name: Test .NET binding
-        if: ${{ matrix.mise_env != 'macos-arm64-egl' }}
+        if: ${{ matrix.mise_env != 'macos-arm64-egl' && !startsWith(matrix.mise_env, 'android-') }}
         run: mise run //bindings/dotnet:test
       # linux-arm64: no Kotlin/Native host — https://youtrack.jetbrains.com/issue/KT-36871
       # windows: no MSVC target — https://youtrack.jetbrains.com/issue/SKIKO-446
       - name: Test Kotlin/Native binding
-        if: ${{ matrix.mise_env != 'macos-arm64-egl' && !startsWith(matrix.mise_env, 'windows-') && !startsWith(matrix.mise_env, 'linux-arm64-') }}
+        if: ${{ matrix.mise_env != 'macos-arm64-egl' && !startsWith(matrix.mise_env, 'android-') && !startsWith(matrix.mise_env, 'windows-') && !startsWith(matrix.mise_env, 'linux-arm64-') }}
         run: mise run //bindings/kotlin-native:build
       - name: Test Rust binding
-        if: ${{ matrix.mise_env != 'macos-arm64-egl' }}
+        if: ${{ matrix.mise_env != 'macos-arm64-egl' && !startsWith(matrix.mise_env, 'android-') }}
         run: mise run //bindings/rust:ci
       - name: Test Swift binding
         if: ${{ matrix.mise_env != 'macos-arm64-egl' && startsWith(matrix.mise_env, 'macos-') }}
         run: mise run //bindings/swift:test
       - name: Test Zig binding
+        if: ${{ !startsWith(matrix.mise_env, 'android-') }}
         run: mise run //bindings/zig:ci
 
       - name: Build LWJGL example
-        if: ${{ contains(matrix.mise_env, 'vulkan') }}
+        if: ${{ contains(matrix.mise_env, 'vulkan') && !startsWith(matrix.mise_env, 'android-') }}
         run: mise run //examples/lwjgl-map:build
       - name: Test Rust example
-        if: ${{ matrix.mise_env != 'macos-arm64-egl' }}
+        if: ${{ matrix.mise_env != 'macos-arm64-egl' && !startsWith(matrix.mise_env, 'android-') }}
         run: mise run //examples/rust-map:ci
       - name: Build Swift example
         if: ${{ matrix.mise_env == 'macos-arm64-metal' }}
         run: mise run //examples/swift-map:build
       - name: Build Zig example
-        if: ${{ matrix.mise_env != 'macos-arm64-egl' }}
+        if: ${{ matrix.mise_env != 'macos-arm64-egl' && !startsWith(matrix.mise_env, 'android-') }}
         run: mise run //examples/zig-map:build
       - name: Run Zig readback example
+        if: ${{ !startsWith(matrix.mise_env, 'android-') }}
         run: mise run //examples/zig-readback:run
 
   required:
