+++ cmake/mln_rust.cmake
@@ -9,7 +9,7 @@ function(mln_link_rust_platform target)
       FATAL_ERROR "CARGO_BUILD_TARGET must be set for Rust platform builds")
   endif()
 
-  set(rust_manifest "${PROJECT_SOURCE_DIR}/platform/rust/Cargo.toml")
+  set(rust_manifest "${PROJECT_SOURCE_DIR}/src/platform/rust/Cargo.toml")
   set(rust_library
       "${PROJECT_SOURCE_DIR}/target/${rust_target}/release/libmaplibre_native_platform.a")
 
@@ -41,7 +41,9 @@ function(mln_link_rust_platform target)
       "${rust_target}"
       --release
     DEPENDS
-      "${rust_manifest}" "${PROJECT_SOURCE_DIR}/platform/rust/src/lib.rs"
+      "${rust_manifest}" "${PROJECT_SOURCE_DIR}/src/platform/rust/src/lib.rs"
+      "${PROJECT_SOURCE_DIR}/src/platform/rust/src/http.rs"
+      "${PROJECT_SOURCE_DIR}/src/platform/rust/src/image.rs"
       "${PROJECT_SOURCE_DIR}/Cargo.toml" "${PROJECT_SOURCE_DIR}/Cargo.lock"
     WORKING_DIRECTORY "${PROJECT_SOURCE_DIR}"
     VERBATIM)
