+++ src/platform/rust/image.cpp
