+++ src/platform/rust/src/lib.rs
@@ -1,0 +1,2 @@
+mod http;
+mod image;
